+++ src/js/validator/numeric.js
@@ -1,5 +1,10 @@
 (function($) {
     $.fn.bootstrapValidator.validators.numeric = {
+        html5Attributes: {
+            message: 'message',
+            separator: 'separator'
+        },
+
         /**
          * Validate decimal number
          *
@@ -7,12 +12,17 @@
          * @param {jQuery} $field Field element
          * @param {Object} options Consist of key:
          * - message: The invalid message
+         * - separator: The decimal separator. Can be "." (default), ","
          * @returns {Boolean}
          */
         validate: function(validator, $field, options) {
             var value = $field.val();
             if (value == '') {
                 return true;
+            }
+            var separator = options.separator || '.';
+            if (separator != '.') {
+                value = value.replace(separator, '.');
             }
 
             return !isNaN(parseFloat(value)) && isFinite(value);